Create Snakebody Class
The test for function of this class is required. This class is not complete. It use double linked list to store the coordinate of snake body. It is very easy to add one unit or delete the last unit for it is double linked list. It is a little bit complex when it comes to check whether game is over or not.