Update Game Map
Here, I changed the variable type of the gamemap. The gamemap requires reset each time snake moves. If redefine the map each time, it takes a lot of time to do so. If  I set a basemap, I could not asign the basemap to gamemap with only one row of code. So, I change the char type to vector<string> type. I would like to check whether vector<string> type assignment is more time-saving.