Update main.cpp
A lot of updates here. First, move and grow_and_move function are updated to make the snake move correctly. Secondly, the input function. It supports that snake moves when there is no keyboard input in a period of time. Third one is a part of run function. When the snake eat the food, it will grow and the food will regenerate. However, here is a little bug which will be fixed soon. The fourth one is adding a function which set the invaild the input whoes direction is opposite to heading direction. The last one is print function. It allows visualize the game.