Update main.cpp
In this update, the structure of the check() function and print() function is updated.  There are two check flags: is_grow and is_over. These two flags are checked when the map is forming (adding body and food to the map). In this case, time cpmlexity is reduced. 

The food generation is updated. The generated location is not possible to be the board of the map. In addition, score and level are added.